Fix Header View not displaying in the right location.
The cause of the problem is that decor view's children handle fitting system windows, which means that any View's added after them do not receive a fitSystemWindows() call. The fix was to insert a layer in between the decor view and it's existing children, meaning I can inset the header view properly. #win

Fix #5.